refactor: rename bridge circuit core to bridge circuit common
