+++ bridge-circuit-host/src/lib.rs
@@ -3,7 +3,7 @@ use alloy_primitives::U256;
 use alloy_rpc_client::RpcClient;
 use alloy_rpc_types::EIP1186AccountProofResponse;
 use anyhow::bail;
-use circuits_lib::bridge_circuit_core::structs::{LightClientProof, StorageProof};
+use circuits_lib::bridge_circuit_common::structs::{LightClientProof, StorageProof};
 use hex::decode;
 use risc0_zkvm::{InnerReceipt, Receipt};
 use serde_json::json;
