+++ bridge-circuit-host/src/structs.rs
@@ -1,6 +1,6 @@
 use ark_bn254::Bn254;
 use ark_ff::PrimeField;
-use circuits_lib::bridge_circuit_core::{
+use circuits_lib::bridge_circuit_common::{
     structs::{LightClientProof, StorageProof},
     winternitz::WinternitzHandler,
 };
