+++ circuits-lib/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod bridge_circuit;
-pub mod bridge_circuit_core;
+pub mod bridge_circuit_common;
 pub mod common;
 pub mod work_only;
