+++ circuits-lib/src/work_only/mod.rs
@@ -1,6 +1,6 @@
-use crate::{bridge_circuit_core, common::zkvm::ZkvmGuest};
+use crate::{bridge_circuit_common, common::zkvm::ZkvmGuest};
 
-use bridge_circuit_core::structs::{WorkOnlyCircuitInput, WorkOnlyCircuitOutput};
+use bridge_circuit_common::structs::{WorkOnlyCircuitInput, WorkOnlyCircuitOutput};
 use crypto_bigint::{Encoding, U128, U256};
 use risc0_zkvm::guest::env;
